Consistent naming for inline ADD files
I thought the cache kept getting invalidated because of the mtime or
ctime or something of these inlined files being added to the context.

Turns out it's just because I kept giving them a unique name everytime!

So now I use an md5 hash so they change when the content changes, and I
use the dest to try and ensure that if we inline two things that are
somehow different but with the same md5sum, we don't have a conflict